docs(nx-dev): switch to using catch-all route for technologies paths
