--- astro-docs/src/pages/api/plugins/[plugin]/migrations.astro
@@ -1,35 +1,0 @@
----
-import StarlightPage from '@astrojs/starlight/components/StarlightPage.astro';
-import { getCollection, render } from 'astro:content';
-
-export async function getStaticPaths() {
-  const pluginDocs = await getCollection(
-    'plugin-docs',
-    (entry) => entry.data.docType === 'migrations'
-  );
-
-  // Generate paths for each plugin
-  return pluginDocs.map((p) => ({
-    params: { plugin: p.data.pluginName },
-    props: { doc: p },
-  }));
-}
-
-const { doc } = Astro.props;
-const pluginName = doc.data.pluginName;
-
-if (!doc.body) {
-  throw new Error(`No content found for plugin: ${pluginName}`);
-}
-const {Content, headings } = await render(doc);
----
-
-<StarlightPage
-  frontmatter={{
-    title: `@nx/${pluginName} - migrations`,
-    description: `@nx/${pluginName} plugin migrations`,
-  }}
-  headings={headings || []}
->
-    <Content />
-</StarlightPage>
