+++ astro-docs/src/pages/references/devkit/ngcli_adapter/index.astro
@@ -2,7 +2,7 @@
 import { getEntry, render, getCollection } from 'astro:content';
 import StarlightPage from '@astrojs/starlight/components/StarlightPage.astro';
 
-const doc = await getEntry('devkit-docs', 'ngcli_adapter-overview');
+const doc = await getEntry('nx-reference-packages', 'ngcli_adapter-overview');
 
 if (!doc) {
   throw new Error('angular cli adapter overview doc not found');
